requirements.txt
appdirs==1.4.4
APScheduler==3.10.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dash==2.9.2
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.1
Flask==2.2.3
fonttools==4.39.3
frozendict==2.3.7
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
multitasking==0.0.11
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
plotly==5.14.1
pycparser==2.21
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-futures==1.0.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
yahooquery==2.3.1
yfinance==0.2.14
zipp==3.15.0
zope.interface==6.0
